# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/schedules:
    get:
      tags:
        - Schedules
      summary: List Schedules
      description: Get a paginated list of Schedules with optional filtering and sorting
      operationId: listSchedules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Schedules
      summary: Create Schedule
      description: Create a new Schedule
      operationId: createSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/schedule_employees:
    get:
      tags:
        - ScheduleEmployees
      summary: List ScheduleEmployees
      description: Get a paginated list of ScheduleEmployees with optional filtering and sorting
      operationId: listScheduleEmployees
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleEmployeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ScheduleEmployees
      summary: Create ScheduleEmployee
      description: Create a new ScheduleEmployee
      operationId: createScheduleEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleEmployeeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleEmployee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Schedule:
      type: object
      required:
        - id
        - name
        - is_default
        - order_number
        - time_in
        - time_out
        - is_override_holiday
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        is_default:
          type: boolean
        order_number:
          type: integer
          format: int32
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        is_override_holiday:
          type: boolean
        metadata:
          type: object
...
    CreateScheduleRequest:
      type: object
      required:
        - name
        - is_default
        - order_number
        - time_in
        - time_out
        - is_override_holiday
        - metadata
      properties:
        name:
          type: string
        is_default:
          type: boolean
        order_number:
          type: integer
          format: int32
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        is_override_holiday:
          type: boolean
        metadata:
          type: object
...
    ScheduleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ScheduleEmployee:
      type: object
      required:
        - id
        - employee_id
        - schedule_date
        - order_number
        - time_in
        - time_out
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        schedule_date:
          type: string
          format: date
        order_number:
          type: integer
          format: int32
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        metadata:
          type: object
...
    CreateScheduleEmployeeRequest:
      type: object
      required:
        - employee_id
        - schedule_date
        - order_number
        - time_in
        - time_out
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        schedule_date:
          type: string
          format: date
        order_number:
          type: integer
          format: int32
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        metadata:
          type: object
...
    ScheduleEmployeeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleEmployee'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []