# src/exports/types.rs
//! Public types for Schedule module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate, NaiveTime};
use crate::domain::entity::*;

// ============================================================================
// SCHEDULE TYPES
// ============================================================================

/// Type-safe ID for Schedule
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ScheduleId(pub Uuid);

impl ScheduleId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ScheduleId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ScheduleId> for Uuid {
    fn from(id: ScheduleId) -> Self {
        id.0
    }
}

/// Data transfer object for Schedule
///
/// This is the public representation of Schedule for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleDto {
    pub id: ScheduleId,
    pub name: String,
    pub is_default: bool,
    pub order_number: i32,
    pub start_date: Option<NaiveDate>,
    pub end_date: Option<NaiveDate>,
    pub time_in: NaiveTime,
    pub time_out: NaiveTime,
    pub is_override_holiday: bool,
    pub metadata: serde_json::Value,
}

/// Summary view of Schedule for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleSummary {
    pub id: ScheduleId,
    pub name: String,
}

/// Reference to Schedule for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleRef {
    pub id: ScheduleId,
}

// ============================================================================
// SCHEDULEEMPLOYEE TYPES
// ============================================================================

/// Type-safe ID for ScheduleEmployee
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ScheduleEmployeeId(pub Uuid);

impl ScheduleEmployeeId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ScheduleEmployeeId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ScheduleEmployeeId> for Uuid {
    fn from(id: ScheduleEmployeeId) -> Self {
        id.0
    }
}

/// Data transfer object for ScheduleEmployee
///
/// This is the public representation of ScheduleEmployee for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleEmployeeDto {
    pub id: ScheduleEmployeeId,
    pub employee_id: Uuid,
    pub schedule_date: NaiveDate,
    pub order_number: i32,
    pub time_in: NaiveTime,
    pub time_out: NaiveTime,
    pub metadata: serde_json::Value,
}

/// Summary view of ScheduleEmployee for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleEmployeeSummary {
    pub id: ScheduleEmployeeId,
}

/// Reference to ScheduleEmployee for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleEmployeeRef {
    pub id: ScheduleEmployeeId,
}

// ============================================================================
// SCHEDULEORGANIZATION TYPES
// ============================================================================

/// Type-safe ID for ScheduleOrganization
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ScheduleOrganizationId(pub Uuid);

impl ScheduleOrganizationId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ScheduleOrganizationId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ScheduleOrganizationId> for Uuid {
    fn from(id: ScheduleOrganizationId) -> Self {
        id.0
    }
}

/// Data transfer object for ScheduleOrganization
///
/// This is the public representation of ScheduleOrganization for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleOrganizationDto {
    pub id: ScheduleOrganizationId,
    pub structure_id: Option<Uuid>,
    pub name: Option<String>,
    pub order_number: i32,
    pub start_date: Option<NaiveDate>,
    pub end_date: Option<NaiveDate>,
    pub time_in: NaiveTime,
    pub time_out: NaiveTime,
    pub metadata: serde_json::Value,
}

/// Summary view of ScheduleOrganization for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleOrganizationSummary {
    pub id: ScheduleOrganizationId,
    pub name: Option<String>,
}

/// Reference to ScheduleOrganization for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleOrganizationRef {
    pub id: ScheduleOrganizationId,
}

// ============================================================================
// SCHEDULEWEEKDAY TYPES
// ============================================================================

/// Type-safe ID for ScheduleWeekday
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ScheduleWeekdayId(pub Uuid);

impl ScheduleWeekdayId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ScheduleWeekdayId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ScheduleWeekdayId> for Uuid {
    fn from(id: ScheduleWeekdayId) -> Self {
        id.0
    }
}

/// Data transfer object for ScheduleWeekday
///
/// This is the public representation of ScheduleWeekday for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleWeekdayDto {
    pub id: ScheduleWeekdayId,
    pub schedule_id: Uuid,
    pub sun: WeekdayType,
    pub mon: WeekdayType,
    pub tue: WeekdayType,
    pub wed: WeekdayType,
    pub thu: WeekdayType,
    pub fri: WeekdayType,
    pub sat: WeekdayType,
    pub metadata: serde_json::Value,
}

/// Summary view of ScheduleWeekday for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleWeekdaySummary {
    pub id: ScheduleWeekdayId,
}

/// Reference to ScheduleWeekday for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ScheduleWeekdayRef {
    pub id: ScheduleWeekdayId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>